# Make equality checks faster and fix hash