--- a/src/base.jl
+++ b/src/base.jl
@@ -38,7 +38,7 @@ Base.lastindex(s::ShortString) = sizeof(s)
 Base.ncodeunits(s::ShortString) = sizeof(s)
 Base.print(s::ShortString) = print(String(s))
 Base.show(io::IO, str::ShortString) = show(io, String(str))
-Base.sizeof(s::ShortString{T}) where T = Int(s.size_content & size_mask(T))
+Base.sizeof(s::ShortString{T}) where T = Int(s.size_content & (size_mask(s) % UInt))
 
 size_nibbles(::Type{<:Union{UInt16, UInt32, UInt64, UInt128}}) = 1
 size_nibbles(::Type{<:Union{Int16, Int32, Int64, Int128}}) = 1
@@ -46,7 +46,8 @@ size_nibbles(::Type{<:Union{UInt256, UInt512, UInt1024}}) = 2
 size_nibbles(::Type{<:Union{Int256, Int512, Int1024}}) = 2
 size_nibbles(::Type{T}) where T = ceil(log2(sizeof(T))/4)
 
-size_mask(T) = UInt(exp2(4*size_nibbles(T)) - 1)
+size_mask(T) = T(exp2(4*size_nibbles(T)) - 1)
+size_mask(s::ShortString{T}) where T = size_mask(T)
 
 
 # function Base.getindex(s::ShortString, i::Integer)
@@ -59,7 +60,7 @@ size_mask(T) = UInt(exp2(4*size_nibbles(T)) - 1)
 
 Base.collect(s::ShortString) = collect(String(s))
 
-function ==(s::ShortString{S}, b::Union{String, SubString{String}) where S
+function ==(s::ShortString{S}, b::Union{String, SubString{String}}) where S
     ncodeunits(b) == ncodeunits(s) || return false
     return s == ShortString{S}(b)
 end
@@ -73,6 +74,13 @@ end
 function ==(a::ShortString{S}, b::ShortString{S}) where S
     return a.size_content == b.size_content
 end
+function ==(a::ShortString{A}, b::ShortString{B}) where {A,B}
+    ncodeunits(a) == ncodeunits(b) || return false
+    # compare if equal after droppign size bits and
+    # flipping so that the empty bytes are at the start
+    ntoh(a.size_content & ~size_mask(A)) == ntoh(b.size_content & ~size_mask(B))
+end
+
 
 promote_rule(::Type{String}, ::Type{ShortString{S}}) where S = String
 promote_rule(::Type{ShortString{T}}, ::Type{ShortString{S}}) where {T,S} = ShortString{promote_rule(T,S)}