--- a/src/hash.jl
+++ b/src/hash.jl
@@ -2,6 +2,4 @@ export hash
 
 import Base.hash
 
-Base.hash(x::ShortString, args...; kwargs...) = hash(x.size_content, args...; kwargs...)
-
-Base.hash(x::ShortString, h::UInt) = hash(x.size_content, h)
+Base.hash(x::ShortString, h::UInt) = hash(String(x), h)