--- a/test/hash.jl
+++ b/test/hash.jl
@@ -1,5 +1,0 @@
-using ShortStrings: ShortString, hash
-using Test
-
-
-@test hash(ShortString(10)) == hash(UInt(10))