--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -3,7 +3,6 @@ using BitIntegers: UInt256, UInt512, UInt1024, @define_integers
 using Test, Random
 
 include("getindex.jl")
-include("hash.jl")
 
 function basic_test(constructor, max_len)
     @testset "$constructor" begin
@@ -18,6 +17,7 @@ end
 function basic_test(string_type, constructor, max_len)
     r = string_type.(randstring.(1:max_len))
     @test all(constructor.(r) .== r)
+    @test all(hash(constructor.(r)) .== hash(r))
     a = constructor.(r)
     @test fsort(a) |> issorted