solves #15 

With this PR:
```julia
julia> @btime $(ShortString15("abc")) == "abc";
  10.935 ns (0 allocations: 0 bytes)

julia> @btime $(ShortString15("abc")) == "abd";
  11.152 ns (0 allocations: 0 bytes)
```

Without this PR
```julia
julia> @btime $(ShortString15("abc")) == "abc";
  79.701 ns (3 allocations: 224 bytes)

julia> @btime $(ShortString15("abc")) == "abd";
  80.259 ns (3 allocations: 224 bytes)
```

This also correct the `hash` closing #21